minor parsing fix, improve grouping styling
