+++ lib/dotcom/trip_plan/open_trip_planner.ex
@@ -4,7 +4,7 @@ defmodule Dotcom.TripPlan.OpenTripPlanner do
   parses the result.
   """
 
-  alias Dotcom.TripPlan.{InputForm, Leg, NamedPosition, Parser, PersonalDetail}
+  alias Dotcom.TripPlan.{InputForm, NamedPosition, Parser}
 
   alias OpenTripPlannerClient.ItineraryTag.{
     EarliestArrival,
@@ -50,68 +50,5 @@ defmodule Dotcom.TripPlan.OpenTripPlanner do
 
   defp parse({:ok, %OpenTripPlannerClient.Plan{itineraries: itineraries}}) do
     {:ok, Enum.map(itineraries, &Parser.parse/1)}
-  end
-
-  # TODO: Move all of these functions to a more appropriate module.
-
-  @doc """
-  Group itineraries by unique legs.
-
-  A unique leg is defined as a leg that has a unique combination of mode, from, and to.
-  But, this does not include walking legs that are less than 0.2 miles.
-  """
-  def group(itineraries) do
-    itineraries
-    |> Enum.group_by(&unique_legs_to_hash/1)
-    |> Enum.map(&group_departures/1)
-    |> Enum.reject(&Enum.empty?(&1))
-  end
-
-  defp group_departures({_, group}) do
-    group
-    |> Enum.uniq_by(&itinerary_to_hash/1)
-    |> Enum.map(fn group ->
-      %{
-        departure: group.start,
-        legs: group.legs
-      }
-    end)
-  end
-
-  defp combined_leg_to_tuple(%Leg{mode: %PersonalDetail{}} = leg) do
-    unique_leg_to_tuple(leg)
-  end
-
-  defp combined_leg_to_tuple(%Leg{mode: %{route: route}} = leg) do
-    {route.id, leg.from.name, leg.to.name}
-  end
-
-  defp itinerary_to_hash(itinerary) do
-    itinerary
-    |> Map.get(:legs)
-    |> Enum.reject(&short_walking_leg?/1)
-    |> Enum.map(&combined_leg_to_tuple/1)
-    |> :erlang.phash2()
-  end
-
-  defp short_walking_leg?(%Leg{mode: %PersonalDetail{}} = leg) do
-    leg.distance <= 0.2
-  end
-
-  defp short_walking_leg?(_), do: false
-
-  defp unique_legs_to_hash(legs) do
-    legs
-    |> Enum.reject(&short_walking_leg?/1)
-    |> Enum.map(&unique_leg_to_tuple/1)
-    |> :erlang.phash2()
-  end
-
-  defp unique_leg_to_tuple(%Leg{mode: %PersonalDetail{}} = leg) do
-    {:WALK, leg.from.name, leg.to.name}
-  end
-
-  defp unique_leg_to_tuple(%Leg{mode: %{route: route}} = leg) do
-    {Routes.Route.type_atom(route.type), leg.from.name, leg.to.name}
   end
 end
