+++ lib/dotcom_web/components/trip_planner/itinerary_group.ex
@@ -14,34 +14,28 @@ defmodule DotcomWeb.Components.TripPlanner.ItineraryGroup do
   """
   def itinerary_group(assigns) do
     ~H"""
-    <div class="itinerary-group m-3 mt-0 p-3 border-solid border-1 border-indigo-200">
-      <div class="itinerary-group-legs">
-        <ol class="list-decimal">
-          <%= for variation <- @group do %>
-            <li class="itinerary-group-leg">
-              <h6>
-                DEPARTURE TIME
-                <span class="text-small text-slate-500">
-                  <%= format_datetime(variation.departure) %>
-                </span>
-              </h6>
-              <ul class="list-none">
-                <%= for leg <- variation.legs do %>
-                  <li>
-                    <.leg
-                      start_time={leg.start}
-                      end_time={leg.stop}
-                      from={leg.from}
-                      to={leg.to}
-                      mode={leg.mode}
-                    />
-                  </li>
-                <% end %>
-              </ul>
-            </li>
-          <% end %>
-        </ol>
-      </div>
+    <div class="mb-3 p-2 bg-slate-700">
+      <div class="text-slate-100">Group with <%= Enum.count(@group) %> options</div>
+      <%= for variation <- @group do %>
+        <div class="mb-1 p-1 bg-indigo-100">
+          <div class="text-2xl">
+            <%= format_datetime(variation.departure) %> — <%= format_datetime(variation.arrival) %>
+          </div>
+          <ul class="list-none p-0">
+            <%= for leg <- variation.legs do %>
+              <li>
+                <.leg
+                  start_time={leg.start}
+                  end_time={leg.stop}
+                  from={leg.from}
+                  to={leg.to}
+                  mode={leg.mode}
+                />
+              </li>
+            <% end %>
+          </ul>
+        </div>
+      <% end %>
     </div>
     """
   end
